# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: SessionLayer Control Plane API
  version: 0.1.0
  summary: Contract-first REST surface for the SessionLayer Control Plane.
  description: |
    The **SessionLayer Control Plane API** is the contract-first source of truth
    for the platform's REST surface (Design §13, FR-API-1). The UI is a *client* of
    this contract; server interfaces and typed clients are code-generated from this
    document, and CI fails if generated code drifts from the spec. Postgres is the
    single source of truth; the UI and the API share one write path (D11).

    **Complete contract (Session 17).** This document is the complete, frozen
    resource surface: the config resources (`node-policies`, `rules`, `roles` +
    `role-bindings`, `cas`, `capability-defs`, `jit-policies`,
    `breakglass-policies`, `service-accounts`), the runtime resources (`nodes`,
    `sessions`, `locks`, `jit-requests`, `join-tokens`, `otp`, device-flow, `pins`,
    `breakglass`, `recordings`, `audit-events`), and the operational probes. The
    `recordings` replay/export and `audit-events` search operations are frozen here
    and return `501` until Session 18 implements them, so the Dashboard can be built
    in parallel against a finalized typed client.

    **Config vs runtime.** Config resources carry an `origin` provenance label
    (`api|ui|default`) and an optimistic-concurrency `version`; runtime resources
...
    API-only runtime resource (Design §13).

...
    compatibility policy.
...
servers:
...
security:
  - oidcBearer: []
  - clientCredentials: []
  - mtls: []
paths:
...
  /v1/join-tokens:
    get:
      operationId: listJoinTokens
      summary: List active join tokens (admin).
      description: |
        Lists the current unconsumed, unexpired join tokens. The raw token value
        is NEVER returned (only its metadata) — it is shown exactly once at
        issuance. Platform-RBAC gated (`node:enroll`) + audited (Design §8.1;
        FR-JOIN-2).
      tags:
        - join-tokens
      responses:
        '200':
          description: The active join tokens (metadata only, never the raw token).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JoinTokenList'
        '403':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
    post:
      operationId: issueJoinToken
      summary: Issue an agent join token (admin).
      description: |
        Issues a short-lived, single-use, self-destruct join token bound to a
        stable node identity/scope (Design §8.1; FR-JOIN-2/6). The raw token is
        returned exactly ONCE (for out-of-band delivery to the joining Agent);
        only its SHA-256 hash is stored. Because issuance is a pure API
        operation, an autoscaler/config-mgmt re-provisions a token-join agent
        after a full lapse without a human. Platform-RBAC gated (`node:enroll`)
        + audited.
      tags:
        - join-tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IssueJoinTokenRequest'
      responses:
        '201':
          description: A join token was issued (raw value returned once).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssuedJoinToken'
        '400':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
  /v1/join-tokens/{joinTokenId}:
    delete:
      operationId: revokeJoinToken
      summary: Revoke a join token (admin).
      description: |
        Revokes (deletes) an unconsumed join token so it can never be used.
        Idempotent. Revoking an already-consumed token has no effect on the
        identity it produced (revocation of an issued identity is a Lock, §8.1).
        Platform-RBAC gated (`node:enroll`) + audited.
      tags:
        - join-tokens
      parameters:
        - name: joinTokenId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: The join token was revoked (idempotent).
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
...
  /v1/audit-events:
    get:
      operationId: searchAuditEvents
      summary: Search the correlated audit stream (admin).
      description: |
        Searches the correlated audit stream (`runtime.audit_event`; Design §12,
        FR-AUD-8/9) — one append-only stream for SSH-session and web/admin events,
        correlated by id, newest-first. Search dimensions (FR-AUD-8): identity
        (`actor`), subject, action, outcome, node (`nodeId`) and node label
        (`nodeLabel`), session, source IP, capability, access model, and a time
        range; `correlationId` reconstructs one full path (approve → connect → run →
        replay, FR-AUD-9). Cursor-paginated. Read-only — a search never mutates the
        append-only stream and the hash chain stays verifiable. Platform-RBAC gated
        (`audit:read`); results are additionally filtered to the caller's RBAC scope
        (a node-label/user/time-scoped `audit:read` grant returns only in-scope
        events; an unscoped grant sees all).
      tags:
        - audit-events
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - name: actor
          in: query
          required: false
          description: Filter by acting identity.
          schema:
            type: string
        - name: subject
          in: query
          required: false
          description: Filter by the subject the action targeted.
          schema:
            type: string
        - name: action
          in: query
          required: false
          description: Filter by action name (e.g. `lock.create`).
          schema:
            type: string
        - name: outcome
          in: query
          required: false
          description: Filter by outcome.
          schema:
            type: string
        - name: sessionId
          in: query
          required: false
          description: Filter to one session id.
          schema:
            type: string
            format: uuid
        - name: nodeId
          in: query
          required: false
          description: Filter to one node id.
          schema:
            type: string
            format: uuid
        - name: sourceIp
          in: query
          required: false
          description: Filter by source IP.
          schema:
            type: string
        - name: from
          in: query
          required: false
          description: Inclusive lower bound on occurrence time (RFC 3339).
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          required: false
          description: Exclusive upper bound on occurrence time (RFC 3339).
          schema:
            type: string
            format: date-time
        - name: capability
          in: query
          required: false
          description: Filter to events whose capability set includes this capability (FR-AUD-8).
          schema:
            $ref: '#/components/schemas/Capability'
        - name: accessModel
          in: query
          required: false
          description: Filter by the session's access model (FR-AUD-8).
          schema:
            $ref: '#/components/schemas/AccessModel'
        - name: nodeLabel
          in: query
          required: false
          description: |
            Filter by a snapshot node label as `key=value` (FR-AUD-8). Repeatable;
            all supplied labels must match (AND).
          schema:
            type: array
            items:
              type: string
              maxLength: 512
        - name: correlationId
          in: query
          required: false
          description: Filter to one correlation id — the FR-AUD-9 full-path join key.
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: A page of audit events (newest first).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditEventPage'
        '400':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
  /v1/audit-events/{auditEventId}:
    parameters:
      - $ref: '#/components/parameters/AuditEventId'
    get:
      operationId: getAuditEvent
      summary: Get one audit event (admin) — FROZEN, 501 until Session 18.
      description: |
        Returns one audit event by id. Platform-RBAC gated (`audit:read`). FROZEN
        CONTRACT: returns `501` until Session 18.
      tags:
        - audit-events
      responses:
        '200':
          description: The audit event.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditEventResource'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '501':
          $ref: '#/components/responses/NotImplemented'
        default:
          $ref: '#/components/responses/Problem'
components:
  securitySchemes:
    oidcBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        OIDC bearer access via an ID/JWT token validated by the Control Plane
        (Design §5.7, FR-AUTH-17). The ID token is the authentication proof.
    clientCredentials:
      type: oauth2
      description: |
        OAuth 2.0 client-credentials for machine consumers (Design §5.6,
        FR-AUTH-12). `private_key_jwt`/mTLS client authentication is preferred
        over static secrets.
      flows:
        clientCredentials:
          tokenUrl: https://controlplane.sessionlayer.example/oauth2/token
          scopes: {}
    mtls:
      type: mutualTLS
      description: |
        Mutual TLS client-certificate authentication (Design §5.7, FR-AUTH-17),
        also used by the internal CP<->Gateway plane.
  parameters:
    Cursor:
      name: cursor
      in: query
      required: false
      description: |
        Opaque forward-only pagination cursor from a previous page's `nextCursor`.
        Omit for the first page. An unrecognised cursor is a `400`.
      schema:
        type: string
        maxLength: 512
    Limit:
      name: limit
      in: query
      required: false
      description: Maximum items to return in one page (the server clamps to its maximum).
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 200
        default: 50
...
    AuditEventId:
      name: auditEventId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    Problem:
      description: An error occurred. The body is an RFC 9457 problem document.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    NotImplemented:
      description: |
        The operation is frozen in the contract but not yet implemented (Session 18).
        The body is an RFC 9457 problem document with `status` 501.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
  schemas:
    ProblemDetails:
      type: object
      title: RFC 9457 Problem Details
      description: |
        Problem details for HTTP APIs as defined by RFC 9457 (obsoletes RFC 7807).
        Error responses use media type `application/problem+json`.
      properties:
        type:
          type: string
          format: uri-reference
          default: about:blank
          description: A URI reference identifying the problem type.
        title:
          type: string
          description: A short, human-readable summary of the problem type.
        status:
          type: integer
          format: int32
          minimum: 100
          maximum: 599
          description: The HTTP status code.
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence.
        instance:
          type: string
          format: uri-reference
          description: A URI reference identifying the specific occurrence.
      additionalProperties: true
...
    IssueJoinTokenRequest:
      type: object
      title: Issue Join Token Request
      description: Request to mint a single-use agent join token bound to a node.
      required:
        - nodeName
      properties:
        nodeName:
          type: string
          minLength: 1
          maxLength: 253
          description: >-
            The stable node identity the token authorizes an agent to join as
            (FR-JOIN-6). The token is scoped to exactly this node.
        ttlSeconds:
          type: integer
          format: int32
          minimum: 1
          description: >-
            Optional TTL override (seconds). Omitted uses the configured default;
            the value is clamped to the configured maximum. Join tokens are
            short-lived by design.
      additionalProperties: false
    IssuedJoinToken:
      type: object
      title: Issued Join Token
      description: >-
        A freshly issued join token. The `token` field is the raw single-use
        bearer value and is returned exactly ONCE; only its hash is stored.
      required:
        - id
        - token
        - nodeName
        - joinMethod
        - singleUse
        - expiresAt
      properties:
        id:
          type: string
          format: uuid
        token:
          type: string
          description: The raw single-use join token. Shown once; never retrievable again.
        nodeName:
          type: string
        joinMethod:
          type: string
          enum:
            - token
        singleUse:
          type: boolean
        expiresAt:
          type: string
          format: date-time
      additionalProperties: false
    JoinTokenResource:
      type: object
      title: Join Token
      description: Join-token metadata (never the raw token).
      required:
        - id
        - nodeName
        - joinMethod
        - singleUse
        - expiresAt
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        nodeName:
          type: string
        joinMethod:
          type: string
          enum:
            - token
        singleUse:
          type: boolean
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
          description: The admin identity that issued the token.
      additionalProperties: false
    JoinTokenList:
      type: object
      title: Join Token List
      required:
        - joinTokens
      properties:
        joinTokens:
          type: array
          items:
            $ref: '#/components/schemas/JoinTokenResource'
      additionalProperties: false
...
    Capability:
      type: string
      title: SSH Capability
      description: A policy-gated SSH channel capability (Design D14).
      enum:
        - shell
        - exec
        - sftp
        - scp
        - port_forward_local
        - port_forward_remote
        - agent_forward
        - x11
...
    AccessModel:
      type: string
      title: Access Model
      enum:
        - standing
        - jit
        - breakglass
    Selector:
      type: object
      title: Selector
      description: A shape-validated selector object (stored as jsonb).
      additionalProperties: true
...
    AuditEventResource:
      type: object
      title: Audit Event
      description: One row of the correlated append-only audit stream (Design §12).
      required:
        - id
        - occurredAt
        - actor
        - action
        - outcome
      properties:
        id:
          type: string
          format: uuid
        occurredAt:
          type: string
          format: date-time
        actor:
          type: string
        subject:
          type: string
        action:
          type: string
        outcome:
          type: string
        sessionId:
          type: string
          format: uuid
        nodeId:
          type: string
          format: uuid
        correlationId:
          type: string
          format: uuid
        sourceIp:
          type: string
        detail:
          $ref: '#/components/schemas/Selector'
      additionalProperties: false
    AuditEventPage:
      type: object
      title: Audit Event Page
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AuditEventResource'
        nextCursor:
          type: string
      additionalProperties: false